made it possible to find tables in parent scopes
