+++ go/vt/vtgate/semantics/semantic_state.go
@@ -19,7 +19,6 @@ package semantics
 import (
 	"fmt"
 
-	"vitess.io/vitess/go/mysql"
 	"vitess.io/vitess/go/vt/vtgate/vindexes"
 
 	"vitess.io/vitess/go/vt/sqlparser"
@@ -31,18 +30,7 @@ type (
 		exprScope        map[sqlparser.Expr]*scope
 		exprDependencies map[sqlparser.Expr][]*sqlparser.AliasedTableExpr
 	}
-	scope struct {
-		i      int
-		tables map[string]*sqlparser.AliasedTableExpr
-	}
 )
-
-var i = 0
-
-func newScope() *scope {
-	i++
-	return &scope{i: i, tables: map[string]*sqlparser.AliasedTableExpr{}}
-}
 
 func (t *SemTable) scope(expr sqlparser.Expr) *scope {
 	return t.exprScope[expr]
@@ -78,7 +66,7 @@ func newAnalyzer(si schemaInformation) *analyzer {
 func Analyse(statement sqlparser.Statement, si schemaInformation) (*SemTable, error) {
 	analyzer := newAnalyzer(si)
 	// Initial scope
-	analyzer.push(newScope())
+	analyzer.push(newScope(nil))
 	analyzer.analyze(statement)
 	if analyzer.err != nil {
 		return nil, analyzer.err
@@ -112,99 +100,6 @@ func (a *analyzer) analyze(statement sqlparser.Statement) {
 		sqlparser.Rewrite(stmt.Having, a.scopeExprs, a.bindExprs)
 		sqlparser.Rewrite(stmt.Limit, a.scopeExprs, a.bindExprs)
 	}
-}
-
-func (a *analyzer) scopeExprs(cursor *sqlparser.Cursor) bool {
-	n := cursor.Node()
-	current := a.peek()
-	log(n, "%d scopeExprs %T", current.i, n)
-	switch expr := n.(type) {
-	case *sqlparser.Subquery:
-		a.exprScope[expr] = current
-		a.push(newScope())
-		a.analyze(expr.Select)
-		a.pop()
-		return false
-	case sqlparser.Expr:
-		a.exprScope[expr] = current
-	}
-	return true
-}
-
-func (a *analyzer) bindExprs(cursor *sqlparser.Cursor) bool {
-	n := cursor.Node()
-	current := a.peek()
-	log(n, "%d bindExprs %T", current.i, n)
-	switch expr := n.(type) {
-	case *sqlparser.ColName:
-		if expr.Qualifier.IsEmpty() {
-			// try to guess which table this column belongs to
-			a.resolveUnQualifiedColumn(current, expr)
-			return true
-		}
-
-		a.err = a.resolveQualifiedColumn(expr, current)
-	case *sqlparser.BinaryExpr:
-		a.exprDeps[expr] = append(a.exprDeps[expr.Left], a.exprDeps[expr.Right]...)
-	}
-
-	return a.err == nil
-}
-
-func (a *analyzer) resolveQualifiedColumn(expr *sqlparser.ColName, current *scope) error {
-	qualifier := expr.Qualifier.Name.String()
-	tableExpr, found := current.tables[qualifier]
-	if found {
-		a.exprDeps[expr] = []*sqlparser.AliasedTableExpr{tableExpr}
-		return nil
-	}
-
-	return mysql.NewSQLError(mysql.ERBadFieldError, mysql.SSBadFieldError, "Unknown column '%s'", sqlparser.String(expr))
-}
-
-func (a *analyzer) resolveUnQualifiedColumn(current *scope, expr *sqlparser.ColName) {
-	if len(current.tables) == 1 {
-		for _, tableExpr := range current.tables {
-
-			a.exprDeps[expr] = []*sqlparser.AliasedTableExpr{tableExpr}
-		}
-	}
-}
-
-func (a *analyzer) analyzeTableExprs(tablExprs sqlparser.TableExprs) {
-	for _, tableExpr := range tablExprs {
-		a.analyzeTableExpr(tableExpr)
-	}
-}
-
-func (a *analyzer) analyzeTableExpr(tableExpr sqlparser.TableExpr) bool {
-	log(tableExpr, "analyzeTableExpr %T", tableExpr)
-	switch table := tableExpr.(type) {
-	case *sqlparser.AliasedTableExpr:
-		expr := table.Expr
-		switch t := expr.(type) {
-		case *sqlparser.DerivedTable:
-
-			a.push(newScope())
-			a.analyze(t.Select)
-			a.pop()
-			scope := a.peek()
-			scope.tables[table.As.String()] = table
-		case sqlparser.TableName:
-			scope := a.peek()
-			if table.As.IsEmpty() {
-				scope.tables[t.Name.String()] = table
-			} else {
-				scope.tables[table.As.String()] = table
-			}
-		}
-	case *sqlparser.JoinTableExpr:
-		a.analyzeTableExpr(table.LeftExpr)
-		a.analyzeTableExpr(table.RightExpr)
-	case *sqlparser.ParenTableExpr:
-		a.analyzeTableExprs(table.Exprs)
-	}
-	return true
 }
 
 func (a *analyzer) push(s *scope) {
